move historic visits into residents controller
